Move isoutil into the isoeditor package
isoeditor is and will likely continue to be the only user of these
functions. It doesn't really make sense to have a separate package for
them.
