+++ internal/isoeditor/isoutil_test.go
@@ -1,4 +1,4 @@
-package isoutil
+package isoeditor
 
 import (
 	"io/ioutil"
